Fixed discount issue in single customer query, separated queries and mutations.
