+++ src/graphql/mutations/product.js
@@ -1,0 +1,9 @@
+import { gql } from 'apollo-boost';
+
+export const CREATE_PRODUCT_MUTATION = gql`
+mutation createProduct($data: ProductCreateInput!) {
+    createProduct(data: $data){
+        name
+    }
+}
+`;
