+++ src/graphql/queries/product.js
@@ -1,0 +1,11 @@
+import { gql } from 'apollo-boost';
+
+export const PRODUCTS_QUERY = gql`
+    query ProductQuery {
+        products {
+            id
+            name
+            price
+        }
+    }
+`;
